// types/teacherProfileTypes/SocialLink/types.ts
export interface UpdateSocialLinks {
  facebookUrl?: string;
  youtubeUrl?: string;
  tiktokUrl?: string;
  instagramUrl?: string;
  websiteUrl?: string;
}

export interface SocialLinks {

  facebookUrl: string;

  youtubeUrl: string;

  tiktokUrl: string;

  instagramUrl: string;

  websiteUrl: string;
  
}